fix(setup.py/build_all_and_release.yml): set release version based on env var
useful as when making release candidates, the setup.py need not be hardcoded to same version as the tagged commit as we'll use the tagged commit name to identify the filename on upload. Or to put it another way, for cx_freeze and windows the filename has the setup.py release name built in, so when we do the upload, we need a handle on it, and by using an environment variable we can ensure even on release candidates this is consistent. Yes I know this is all not ideal; would be good to refactor the logic here to be a single simple installer, with a single consistent way of naming files across os-es.