feat: pin cxfreeze to 7.2.0 and switch to pyproject.toml for linux installs
It seems we were using setup.py for linux installs (maybe ignoring the pyproject.toml). I think what happened it that cx freeze got upgraded and as we hadn't pinned the version, we actually switched building behaviour. To be honest before this commit, the linux release .deb file still works  - but the issue I'm seeing is it's not baking in  library properly, so when trying to bake this in I realised the pypyroject.toml was being ignored, hence this change